refactor: 테스트 코드에 util 내 crypto 패키징 적용
