+++ src/test/java/com/coupang/marketplace/user/service/SessionLoginServiceTest.java
@@ -20,7 +20,7 @@ import com.coupang.marketplace.user.controller.SignInRequestDto;
 import com.coupang.marketplace.user.domain.User;
 import com.coupang.marketplace.fixture.UserFixture.*;
 import com.coupang.marketplace.user.repository.UserRepository;
-import com.coupang.marketplace.global.util.Encryptor;
+import com.coupang.marketplace.global.util.crypto.Encryptor;
 
 @ExtendWith(MockitoExtension.class)
 public class SessionLoginServiceTest {
