+++ src/test/java/com/coupang/marketplace/user/service/UserServiceTest.java
@@ -3,7 +3,7 @@ package com.coupang.marketplace.user.service;
 import com.coupang.marketplace.user.controller.SignUpRequestDto;
 import com.coupang.marketplace.user.domain.User;
 import com.coupang.marketplace.user.repository.UserRepository;
-import com.coupang.marketplace.global.util.Encryptor;
+import com.coupang.marketplace.global.util.crypto.Encryptor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
